requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
babel==2.16.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iso8601==1.1.0
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mysql-connector-python==9.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pdfkit==1.0.0
pendulum==3.0.0
pillow==11.0.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.1
rsa==4.9
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
time-machine==2.15.0
tortoise-orm==0.21.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
